# app/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RSS feed aggregator with passkey authentication.",
        "title": "GKFeed API",
        "contact": {
            "name": "GKFeed",
            "url": "https://github.com/gkfeed/api-go"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/auth/credentials": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all registered passkeys for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "List passkeys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "created_at": {
                                                "type": "string"
                                            },
                                            "id": {
                                                "type": "string"
                                            },
                                            "last_used_at": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/auth/credentials/{id}": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a specific passkey credential for the authenticated user.",
                "tags": [
                    "auth"
                ],
                "summary": "Delete passkey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Credential ID (base64url)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/auth/login/begin": {
            "post": {
                "description": "Starts discoverable WebAuthn login, returns assertion options for the browser.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Begin passkey login",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/auth/login/finish": {
            "post": {
                "description": "Completes WebAuthn authentication, returns a JWT access/refresh token pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Finish passkey login",
                "parameters": [
                    {
                        "description": "WebAuthn assertion response",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "credential_id": {
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidates all refresh tokens for the authenticated user.",
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Exchanges a valid refresh token for a new access/refresh token pair. Old refresh token is invalidated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh access token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.refreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/auth/register/begin": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Starts WebAuthn registration, returns creation options for the browser.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Begin passkey registration",
                "parameters": [
                    {
                        "description": "Optional name for the passkey",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.registerBeginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/auth/register/finish": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Completes WebAuthn registration with the attestation response from the browser.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Finish passkey registration",
                "parameters": [
                    {
                        "description": "WebAuthn attestation response",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_handlers.refreshRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.registerBeginRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}